requirements.txt
matplotlib==3.4.3
mlxtend==0.18.0
numpy==1.21.2
pandas==1.2.5
python_dateutil==2.8.2
scikit_learn==1.0.1
scipy==1.7.1
xgboost==1.4.2
